mod: update company member controller

+++ app/Http/Controllers/Api/CompanyMemberController.php
@@ -10,8 +10,10 @@ use App\Models\CompanyMember;
 use App\Http\Requests\StoreCompanyMemberRequest;
 use App\Http\Requests\UpdateCompanyMemberRequest;
 use App\Repositories\CompanyMember\ICompanyMemberRepository;
+use App\Repositories\User\IUserRepository;
 use Exception;
 use Illuminate\Auth\Access\AuthorizationException;
+use Gate;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class CompanyMemberController extends Controller
@@ -26,18 +28,51 @@ class CompanyMemberController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(CompanyMember $companyMember)
+    public function index(Company $company)
     {
-        $members = $this->companyMemberRepository->paginateByCompany($companyMember->id);
-        return $this->sendResponse("Get Company Member List", 200, new GenericCollection($members, CompanyMemberResource::class));
+        try {
+            Gate::authorize('viewAny', CompanyMember::class);
+            $members = $this->companyMemberRepository->paginateByCompany($company->id);
+            return $this->sendResponse("Get Company Member List", 200, new GenericCollection($members, CompanyMemberResource::class));
+        } catch (AuthorizationException $e) {
+            return $this->sendUnauthorized("You do not have permission to do this action");
+        } catch (Exception $ex) {
+            return $this->sendInternalError("Error", $ex);
+        }
+    }
+
+    /**
+     * Display a listing for selection
+     */
+    public function getSelectList()
+    {
+        try {
+            Gate::authorize('view', CompanyMember::class);
+            $members = $this->companyMemberRepository->getMembersNotInCompany();
+            return $this->sendResponse("Get Select List", 200, CompanyMemberResource::collection($members));
+        } catch (AuthorizationException $e) {
+            return $this->sendUnauthorized("You do not have permission to do this action");
+        } catch (Exception $ex) {
+            return $this->sendInternalError("Error", $ex);
+        }
+
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Company $company, StoreCompanyMemberRequest $request)
+    public function store(StoreCompanyMemberRequest $request)
     {
-
+        try {
+            Gate::authorize('create', CompanyMember::class);
+            $data = $request->validated();
+            $result = $this->companyMemberRepository->create($data);
+            return $this->sendResponse("Company Address Created", 200, new CompanyMemberResource($result));
+        } catch (AuthorizationException $e) {
+            return $this->sendUnauthorized("You do not have permission to do this action");
+        } catch (Exception $ex) {
+            return $this->sendInternalError("Error", $ex);
+        }
     }
 
     /**
@@ -60,7 +95,18 @@ class CompanyMemberController extends Controller
      */
     public function update(UpdateCompanyMemberRequest $request, CompanyMember $companyMember)
     {
-        //
+        try {
+            Gate::authorize('update', $companyMember);
+            $data = $request->validated();
+            $result = $this->companyMemberRepository->update($companyMember->id, $data);
+            return $this->sendResponse("Company Member Updated", 200, new CompanyMemberResource($result));
+        } catch (AuthorizationException $e) {
+            return $this->sendUnauthorized("You do not have permission to do this action");
+        } catch (ModelNotFoundException $ex) {
+            return $this->sendNotFound("Company Member is not exist or already deleted");
+        } catch (Exception $ex) {
+            return $this->sendInternalError("Error", $ex);
+        }
     }
 
     /**
@@ -69,6 +115,7 @@ class CompanyMemberController extends Controller
     public function destroy(CompanyMember $companyMember)
     {
         try {
+            Gate::authorize('delete', $companyMember);
             $this->companyMemberRepository->delete($companyMember->id);
             return $this->sendResponse("Company Member Deleted", 200);
         } catch (AuthorizationException $e) {
